+++ CHANGES.md
@@ -21,6 +21,7 @@ v0.11.22 - TBD
 * WASAPI: Fix an error when stopping the device where it was possible miniaudio would not wait for the device to be drained due to an error with the wait time calculation.
 * WASAPI: Fix a COM related crash with device rerouting.
 * DirectSound: Add support for specifying an explicit window handle for SetCooperativeLevel().
+* ALSA: Fix a bug where a playback device can fail to start.
 * ALSA: Fix some warnings relating to unhandled return value of `read()`.
 * Web: Fix ScriptProcessorNode path when compiling with `--closure=1`. Note that the Audio Worklets path is not currently working due to the callback specified in `emscripten_create_wasm_audio_worklet_processor_async` never getting fired.
 * Web: Fix an error with the unlocked notification when compiling as C++.